refactor(provider/oracle): oraclebmcs provider is renamed to oracle (#2287)
